requirements.txt
torch~=2.8.0
Flask~=3.1.2
torchvision~=0.23.0
pillow~=11.3.0
mlflow~=3.1.4
mlflow-skinny~=3.1.4
matplotlib~=3.9.4
seaborn
scikit-learn
python-dotenv
pathlib
kaggle